Pulled ShapeFinder related Mock classes into their own file.
