+++ tests/src/ShapeFinderTests.cpp
@@ -10,30 +10,9 @@
 
 #include "TestOverrides.h"
 #include "TestUtils.h"
+#include "TestMocks.h"
 
 namespace HMDT::UnitTests {
-    class GraphicsWorkerMock: public IGraphicsWorker {
-        public:
-            virtual ~GraphicsWorkerMock() = default;
-
-            virtual void writeDebugColor(uint32_t, uint32_t, const Color&) { }
-            virtual void updateCallback(const Rectangle&) { }
-
-            static GraphicsWorkerMock& getInstance() {
-                static GraphicsWorkerMock instance;
-
-                return instance;
-            }
-    };
-
-    class ShapeFinderMock: public ShapeFinder {
-        public:
-            using ShapeFinder::ShapeFinder;
-
-            using ShapeFinder::pass1;
-            using ShapeFinder::outputStage;
-    };
-
     struct InputImageInfo {
         std::filesystem::path path;
         uint32_t num_border_pixel;
